# dict/api/student/index.ts
/* eslint-disable */
// tslint:disable
/**
 * STUDENT REGISTRATION FORM
 * 
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator+.
 * https://github.com/karlvr/openapi-generator-plus
 * Do not edit the class manually.
 */

import { Express } from 'express'
import passport from 'passport'
import * as t from './types'
import * as v from '../../validation'
import { Api } from '../../models'

export default function(app: Express, impl: t.StudentApi) {
	app.delete(
		'/DeleteStudent/:Id',
		function (req, res) {
			try {
				impl.deletestudent(v.parseString('params.Id', req.params['Id'])).then(function (response) {
					if (response.status === 201) {
						let body: any
						try {
							body = v.modelApiApiResponseToJson('response', response.body)
						} catch (error) {
							console.error('Invalid response body in student.deletestudent', error)
							res.status(500)
							res.send()
							return
						}

						res.status(201)
						res.send(body)
						return
					}
					if (response.status === 400) {
						let body: any
						try {
							body = v.modelApiApiResponseToJson('response', response.body)
						} catch (error) {
							console.error('Invalid response body in student.deletestudent', error)
							res.status(500)
							res.send()
							return
						}

						res.status(400)
						res.send(body)
						return
					}
					if (response.status === 409) {
						let body: any
						try {
							body = v.modelApiApiResponseToJson('response', response.body)
						} catch (error) {
							console.error('Invalid response body in student.deletestudent', error)
							res.status(500)
							res.send()
							return
						}

						res.status(409)
						res.send(body)
						return
					}

					console.log('Unsupported response in student.deletestudent', response)
					res.status(500)
					res.send()
				}).catch(function (error) {
					console.error('Unexpected error in student.deletestudent', error.stack || error)
					res.status(500)
					res.send()
				})
			} catch (error) {
				/* Catch validation errors */
				res.status(400)
				res.send(error)
			}
		}
	)

	app.get(
		'/GetAllStudents',
		function (req, res) {
			try {
				impl.getGetAllStudents().then(function (response) {
					if (response.status === 201) {
						let body: any
						try {
							body = v.arrayToJson(v.modelApiStudentToJson)('response', response.body)
						} catch (error) {
							console.error('Invalid response body in student.getGetAllStudents', error)
							res.status(500)
							res.send()
							return
						}

						res.status(201)
						res.send(body)
						return
					}
					if (response.status === 400) {
						let body: any
						try {
							body = v.modelApiApiResponseToJson('response', response.body)
						} catch (error) {
							console.error('Invalid response body in student.getGetAllStudents', error)
							res.status(500)
							res.send()
							return
						}

						res.status(400)
						res.send(body)
						return
					}
					if (response.status === 409) {
						let body: any
						try {
							body = v.modelApiApiResponseToJson('response', response.body)
						} catch (error) {
							console.error('Invalid response body in student.getGetAllStudents', error)
							res.status(500)
							res.send()
							return
						}

						res.status(409)
						res.send(body)
						return
					}

					console.log('Unsupported response in student.getGetAllStudents', response)
					res.status(500)
					res.send()
				}).catch(function (error) {
					console.error('Unexpected error in student.getGetAllStudents', error.stack || error)
					res.status(500)
					res.send()
				})
			} catch (error) {
				/* Catch validation errors */
				res.status(400)
				res.send(error)
			}
		}
	)

	app.get(
		'/GetStudent/:Id',
		function (req, res) {
			try {
				impl.getGetStudent(v.parseString('params.Id', req.params['Id'])).then(function (response) {
					if (response.status === 201) {
						let body: any
						try {
							body = v.modelApiStudentToJson('response', response.body)
						} catch (error) {
							console.error('Invalid response body in student.getGetStudent', error)
							res.status(500)
							res.send()
							return
						}

						res.status(201)
						res.send(body)
						return
					}
					if (response.status === 400) {
						let body: any
						try {
							body = v.modelApiApiResponseToJson('response', response.body)
						} catch (error) {
							console.error('Invalid response body in student.getGetStudent', error)
							res.status(500)
							res.send()
							return
						}

						res.status(400)
						res.send(body)
						return
					}
					if (response.status === 409) {
						let body: any
						try {
							body = v.modelApiApiResponseToJson('response', response.body)
						} catch (error) {
							console.error('Invalid response body in student.getGetStudent', error)
							res.status(500)
							res.send()
							return
						}

						res.status(409)
						res.send(body)
						return
					}

					console.log('Unsupported response in student.getGetStudent', response)
					res.status(500)
					res.send()
				}).catch(function (error) {
					console.error('Unexpected error in student.getGetStudent', error.stack || error)
					res.status(500)
					res.send()
				})
			} catch (error) {
				/* Catch validation errors */
				res.status(400)
				res.send(error)
			}
		}
	)

	app.post(
		'/StudentRegister',
		function (req, res) {
			try {
				function __body() {
					const __contentType = req.get('Content-Type')
					const __mimeType = __contentType ? __contentType.replace(/;.*/, '') : undefined

					if (__mimeType === 'application/json') {
						return v.modelApiStudentFromJson('body', req.body)
					}
					console.error(`Invalid request content type: ${__contentType}`)
					throw new Error(`Invalid request content type: ${__contentType}`)
				}

				impl.studentcreated(__body()).then(function (response) {
					if (response.status === 201) {
						let body: any
						try {
							body = v.modelApiApiResponseToJson('response', response.body)
						} catch (error) {
							console.error('Invalid response body in student.studentcreated', error)
							res.status(500)
							res.send()
							return
						}

						res.status(201)
						res.send(body)
						return
					}
					if (response.status === 400) {
						let body: any
						try {
							body = v.modelApiApiResponseToJson('response', response.body)
						} catch (error) {
							console.error('Invalid response body in student.studentcreated', error)
							res.status(500)
							res.send()
							return
						}

						res.status(400)
						res.send(body)
						return
					}
					if (response.status === 409) {
						let body: any
						try {
							body = v.modelApiApiResponseToJson('response', response.body)
						} catch (error) {
							console.error('Invalid response body in student.studentcreated', error)
							res.status(500)
							res.send()
							return
						}

						res.status(409)
						res.send(body)
						return
					}

					console.log('Unsupported response in student.studentcreated', response)
					res.status(500)
					res.send()
				}).catch(function (error) {
					console.error('Unexpected error in student.studentcreated', error.stack || error)
					res.status(500)
					res.send()
				})
			} catch (error) {
				/* Catch validation errors */
				res.status(400)
				res.send(error)
			}
		}
	)

	app.put(
		'/UpdateStudent/:Id',
		function (req, res) {
			try {
				function __body() {
					const __contentType = req.get('Content-Type')
					const __mimeType = __contentType ? __contentType.replace(/;.*/, '') : undefined

					if (__mimeType === 'application/json') {
						return v.modelApiStudentUpdateFromJson('body', req.body)
					}
					console.error(`Invalid request content type: ${__contentType}`)
					throw new Error(`Invalid request content type: ${__contentType}`)
				}

				impl.updatestudent(v.parseString('params.Id', req.params['Id']), __body()).then(function (response) {
					if (response.status === 201) {
						let body: any
						try {
							body = v.modelApiApiResponseToJson('response', response.body)
						} catch (error) {
							console.error('Invalid response body in student.updatestudent', error)
							res.status(500)
							res.send()
							return
						}

						res.status(201)
						res.send(body)
						return
					}
					if (response.status === 400) {
						let body: any
						try {
							body = v.modelApiApiResponseToJson('response', response.body)
						} catch (error) {
							console.error('Invalid response body in student.updatestudent', error)
							res.status(500)
							res.send()
							return
						}

						res.status(400)
						res.send(body)
						return
					}
					if (response.status === 409) {
						let body: any
						try {
							body = v.modelApiApiResponseToJson('response', response.body)
						} catch (error) {
							console.error('Invalid response body in student.updatestudent', error)
							res.status(500)
							res.send()
							return
						}

						res.status(409)
						res.send(body)
						return
					}

					console.log('Unsupported response in student.updatestudent', response)
					res.status(500)
					res.send()
				}).catch(function (error) {
					console.error('Unexpected error in student.updatestudent', error.stack || error)
					res.status(500)
					res.send()
				})
			} catch (error) {
				/* Catch validation errors */
				res.status(400)
				res.send(error)
			}
		}
	)

}