# dict/models.ts
/* eslint-disable */
// tslint:disable
/**
 * STUDENT REGISTRATION FORM
 * 
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator+.
 * https://github.com/karlvr/openapi-generator-plus
 * Do not edit the class manually.
 */

export namespace Api {
	/**
	 * <p>ResponseInStatusCode</p>
	 */
	export interface ApiResponse {
		/**
		 * @type {number}
		 * @memberof ApiResponse
		 */
		code?: number;
		'type'?: string;
		message: string;
	}

	export interface Student {
		'Id': string;
		rollnumber: string;
		email: string;
		password: string;
		isExist?: boolean;
	}

	export interface StudentUpdate {
		rollnumber?: string;
		email?: string;
		password?: string;
	}

}